✨ Feature: Create IconButton component with TypeScript definitions and styled components
