+++ src/components/IDE/SideBar/IconButton/IconButton.d.ts
@@ -1,0 +1,7 @@
+import { ReactNode } from 'react';
+
+export interface IconButtonProps {
+	$isActive: boolean;
+	onClick: () => void;
+	icon: ReactNode;
+}
